# Documentation/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Boxinator",
    "version": "v1"
  },
  "paths": {
...
    "/shipments": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentReadDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentReadDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentReadDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shipments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/shipments/guest": {
      "post": {
        "tags": [
          "Shipments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentGuestCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentGuestCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentGuestCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              }
            }
          }
        }
      }
    },
    "/shipments/{shipmentId}": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentEditDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentEditDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentEditDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentReadDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/shipments/customer/{userId}": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentReadDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentReadDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentReadDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/shipments/log/{shipmentId}": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentStatusLogReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentStatusLogReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentStatusLogReadDTO"
                }
              }
            }
          }
        }
      }
    },
    "/shipments/cancel/{shipmentId}": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentStatusLogReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentStatusLogReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentStatusLogReadDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "countryId": {
            "type": "integer",
            "format": "int32"
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "accountType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CountryReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "zoneId": {
            "type": "integer",
            "format": "int32"
          },
          "zoneName": {
            "type": "string",
            "nullable": true
          },
          "countryMultiplier": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "StatusReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BoxReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ShipmentStatusLogReadDTO": {
        "type": "object",
        "properties": {
          "shipmentId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/StatusReadDTO"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ShipmentReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "receiverFirstName": {
            "type": "string",
            "nullable": true
          },
          "receiverLastName": {
            "type": "string",
            "nullable": true
          },
          "receiverAddress": {
            "type": "string",
            "nullable": true
          },
          "receiverZipCode": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "sender": {
            "$ref": "#/components/schemas/UserReadDTO"
          },
          "country": {
            "$ref": "#/components/schemas/CountryReadDTO"
          },
          "boxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoxReadDTO"
            },
            "nullable": true
          },
          "shipmentStatusLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShipmentStatusLogReadDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BoxCreateDTO": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ShipmentCreateDTO": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "integer",
            "format": "int32"
          },
          "receiverFirstName": {
            "type": "string",
            "nullable": true
          },
          "receiverLastName": {
            "type": "string",
            "nullable": true
          },
          "receiverZipCode": {
            "type": "string",
            "nullable": true
          },
          "receiverAddress": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "countryId": {
            "type": "integer",
            "format": "int32"
          },
          "boxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoxCreateDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ShipmentGuestCreateDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "receiverFirstName": {
            "type": "string",
            "nullable": true
          },
          "receiverLastName": {
            "type": "string",
            "nullable": true
          },
          "receiverZipCode": {
            "type": "string",
            "nullable": true
          },
          "receiverAddress": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "countryId": {
            "type": "integer",
            "format": "int32"
          },
          "boxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoxCreateDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ShipmentEditDTO": {
        "type": "object",
        "properties": {
          "receiverFirstName": {
            "type": "string",
            "nullable": true
          },
          "receiverLastName": {
            "type": "string",
            "nullable": true
          },
          "destinationZipCode": {
            "type": "string",
            "nullable": true
          },
          "destinationAddress": {
            "type": "string",
            "nullable": true
          },
          "destinationCountryId": {
            "type": "integer",
            "format": "int32"
          },
          "senderEmail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}